add isOverRocm3d

+++ driver/adaptiveavgpool_driver.hpp
@@ -56,12 +56,12 @@ public:
         data_type = miopen_type<Tgpu>{};
     }
 
-    std::vector<int> ComputeStrides(std::vector<int> input);
+    std::vector<size_t> ComputeStrides(std::vector<size_t> input);
     int AddCmdLineArgs() override;
     int ParseCmdLineArgs(int argc, char* argv[]) override;
     InputFlags& GetInputFlags() override { return inflags; }
 
-    std::vector<int> GetInputTensorDimsFromCmd(const char* param);
+    std::vector<size_t> GetInputTensorDimsFromCmd(const char* param);
     int GetandSetData() override;
 
     int AllocateBuffersAndCopy() override;
@@ -107,8 +107,8 @@ private:
 
     size_t N = 1, C = 1, D = 1, H = 1, W = 1, OD = 1, OH = 1, OW = 1;
 
-    std::vector<int> in_dim;
-    std::vector<int> out_dim;
+    std::vector<size_t> in_dim;
+    std::vector<size_t> out_dim;
     bool isContiguous;
 };
 
@@ -126,11 +126,11 @@ int AdaptiveAvgPoolDriver<Tgpu, Tref>::ParseCmdLineArgs(int argc, char* argv[])
 }
 
 template <typename Tgpu, typename Tref>
-std::vector<int> AdaptiveAvgPoolDriver<Tgpu, Tref>::GetInputTensorDimsFromCmd(const char* param)
+std::vector<size_t> AdaptiveAvgPoolDriver<Tgpu, Tref>::GetInputTensorDimsFromCmd(const char* param)
 {
     std::string lengthsStr = inflags.GetValueStr(param);
 
-    std::vector<int> lengths;
+    std::vector<size_t> lengths;
     std::size_t pos = 0;
     std::size_t new_pos;
 
@@ -150,7 +150,7 @@ std::vector<int> AdaptiveAvgPoolDriver<Tgpu, Tref>::GetInputTensorDimsFromCmd(co
     std::string sliceStr = lengthsStr.substr(pos);
     int len              = std::stoi(sliceStr);
 
-    lengths.push_back(len);
+    lengths.push_back(static_cast<size_t>(len));
 
     return (lengths);
 }
@@ -158,9 +158,9 @@ std::vector<int> AdaptiveAvgPoolDriver<Tgpu, Tref>::GetInputTensorDimsFromCmd(co
 template <typename Tgpu, typename Tref>
 int AdaptiveAvgPoolDriver<Tgpu, Tref>::GetandSetData()
 {
-    in_dim                     = GetInputTensorDimsFromCmd("input_dims");
-    std::vector<int> in_stride = ComputeStrides(in_dim);
-    out_dim                    = GetInputTensorDimsFromCmd("output_dims");
+    in_dim                        = GetInputTensorDimsFromCmd("input_dims");
+    std::vector<size_t> in_stride = ComputeStrides(in_dim);
+    out_dim                       = GetInputTensorDimsFromCmd("output_dims");
     if(in_dim.size() != out_dim.size() + 2)
     {
         MIOPEN_THROW(miopenStatusBadParm,
@@ -199,7 +199,7 @@ int AdaptiveAvgPoolDriver<Tgpu, Tref>::GetandSetData()
         out_dim_final.push_back(OH);
         out_dim_final.push_back(OW);
     }
-    std::vector<int> out_grad_stride = ComputeStrides(out_dim_final);
+    std::vector<size_t> out_grad_stride = ComputeStrides(out_dim_final);
     SetTensorNd(inputDesc, in_dim, in_stride, data_type);
     SetTensorNd(outputDesc, out_dim_final, data_type);
     SetTensorNd(outputGradDesc, out_dim_final, out_grad_stride, data_type);
@@ -210,11 +210,11 @@ int AdaptiveAvgPoolDriver<Tgpu, Tref>::GetandSetData()
 
 // Equivalent to: tensor.tranpose(0, -1).contiguous().tranpose(0, -1) incase contiguous = False
 template <typename Tgpu, typename Tref>
-std::vector<int> AdaptiveAvgPoolDriver<Tgpu, Tref>::ComputeStrides(std::vector<int> inputDim)
+std::vector<size_t> AdaptiveAvgPoolDriver<Tgpu, Tref>::ComputeStrides(std::vector<size_t> inputDim)
 {
     if(!isContiguous)
         std::swap(inputDim.front(), inputDim.back());
-    std::vector<int> strides(inputDim.size());
+    std::vector<size_t> strides(inputDim.size());
     strides.back() = 1;
     for(int i = inputDim.size() - 2; i >= 0; --i)
         strides[i] = strides[i + 1] * inputDim[i + 1];
